requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40
scikit-learn
numpy
pandas
matplotlib
seaborn
tqdm

# API serving
fastapi
uvicorn[standard]

# Explainability
captum

# Validation utilities
iterative-stratification
pyyaml
